requirements.txt
appnope==0.1.0
backcall==0.1.0
colorama==0.4.0
decorator==4.3.0
Django==1.11.1
django-encrypted-id==0.3.1
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.4
ptyprocess==0.6.0
pycrypto<2.6
Pygments==2.2.0
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
structlog==18.2.0
traitlets==4.3.2
wcwidth==0.1.7
safety==1.8.4
black==18.9b0
